+++ .changeset/tidy-expo-ios-import.md
@@ -1,0 +1,5 @@
+---
+"@clerk/expo": patch
+---
+
+Fixes iOS development builds across Expo SDK versions by moving shared native bridge APIs into a separate iOS module.
