+++ packages/expo/ios/ClerkUserButtonNativeView.swift
@@ -1,5 +1,6 @@
 import React
 import UIKit
+import ClerkExpoBridge
 
 public class ClerkUserButtonNativeView: ClerkNativeViewHost {
   override func makeHostedController() -> UIViewController? {
